nix: make internal/nix.cmd public (#2473)
Move `internal/nix.cmd` and friends to the top-level nix package and
export it.

The constructor function is now a `Nix.Command` method. New commands use
`Nix.Logger` for debug logs and include default arguments from
`Nix.ExtraArgs`.

As a convenience, the package-level `Command` function calls
`Default.Command` (which uses the default Nix installation found in
$PATH).
