+++ internal/nix/build.go
@@ -24,7 +24,7 @@ func Build(ctx context.Context, args *BuildArgs, installables ...string) error {
 	FixInstallableArgs(installables)
 
 	// --impure is required for allowUnfreeEnv/allowInsecureEnv to work.
-	cmd := command("build", "--impure")
+	cmd := Command("build", "--impure")
 	cmd.Args = appendArgs(cmd.Args, args.Flags)
 	cmd.Args = appendArgs(cmd.Args, installables)
 	// Adding extra substituters only here to be conservative, but this could also
